Fix UI pixel aspect, compact scrollbar art, unsigned wire byte
- Add per-game UI_PIXEL_ASPECT and scale font glyphs vertically so
  square authored pixels render square in the normalized UI scene.
- Replace per-part scrollbar image/texcoord with compact
  uiScrollBarImage_t (shared UV crop); cl_scrn.c draws it alongside the
  legacy backdrop path.
- Cast MSG_ReadByte to BYTE when reading buffer sizes so values
  128..255 don't sign-extend (cl_layout/cl_parse/sv_game).
- Clip text areas to their viewport (DRAW_CLIP).
- WoW quest buttons use GameFontNormal gold.
- Document ui-payloads and UI system constraints.

+++ client/cl_layout.c
@@ -301,7 +301,8 @@ LPCUIFRAME SCR_Clear(HANDLE data) {
         if (msg.readcount + sizeof(BYTE) > msg.cursize) {
             break;
         }
-        ent->buffer.size = MSG_ReadByte(&msg);
+        /* Buffer length is an unsigned wire byte; values 128..255 must not sign-extend. */
+        ent->buffer.size = (BYTE)MSG_ReadByte(&msg);
         if (msg.readcount + ent->buffer.size > msg.cursize) {
             break;
         }
